requirements.txt
fastapi==0.104.1
SQLAlchemy==2.0.23
pandas==2.1.2
pg8000==1.30.1
uvicorn==0.23.2
python-dotenv==1.0.0
scikit-learn==1.3.2
pymysql==1.1.0